Elasticsearch Record index configuration
https://mitlibraries.atlassian.net/browse/DIP-167

Most of this configuration was taken from an auto generated index. I've changed some fields to text instead of keyword that I think make sense. Others may want adjustments later. `ignore_above` value probably needs some analysis.

To ensure we always use these rules for creating an index, we check for index existence in the Parse command and create one if it doesn't exist. Otherwise, one is automatically created for us based on the Record struct in go which isn't exactly what we want here.

Follow on work will take place under:
https://mitlibraries.atlassian.net/browse/DIP-177
https://mitlibraries.atlassian.net/browse/DIP-178
https://mitlibraries.atlassian.net/browse/DIP-179
