+++ main.go
@@ -3,15 +3,8 @@ package main
 import (
 	"context"
 	"log"
-	"net/http"
 	"os"
 
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/credentials/ec2rolecreds"
-	"github.com/aws/aws-sdk-go/aws/ec2metadata"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/olivere/elastic"
-	aws "github.com/olivere/elastic/aws/v4"
 	"github.com/urfave/cli"
 )
 
@@ -101,6 +94,16 @@ func main() {
 					if err != nil {
 						return err
 					}
+
+					// check if requested index exists, if not, create it.
+					exists, err := client.IndexExists(index).Do(context.Background())
+					if err != nil {
+						return err
+					}
+					if !exists {
+						createRecordIndex(client, index)
+					}
+
 					es, err := client.BulkProcessor().Name("IngestWorker-1").Do(context.Background())
 					if err != nil {
 						return err
@@ -142,12 +145,7 @@ func main() {
 				if err != nil {
 					return err
 				}
-				ctx := context.Background()
-				_, err = client.CreateIndex(index).Do(ctx)
-				if err != nil {
-					return err
-				}
-				log.Println("Index created")
+				createRecordIndex(client, index)
 				return nil
 			},
 		},
@@ -157,27 +155,4 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-func esClient(url string, index string, v4 bool) (*elastic.Client, error) {
-	var client *http.Client
-	if v4 {
-		sess := session.Must(session.NewSession())
-		creds := credentials.NewChainCredentials([]credentials.Provider{
-			&credentials.EnvProvider{},
-			&credentials.SharedCredentialsProvider{},
-			&ec2rolecreds.EC2RoleProvider{
-				Client: ec2metadata.New(sess),
-			},
-		})
-		client = aws.NewV4SigningClient(creds, "us-east-1")
-	} else {
-		client = http.DefaultClient
-	}
-	return elastic.NewClient(
-		elastic.SetURL(url),
-		elastic.SetSniff(false),
-		elastic.SetHealthcheck(false),
-		elastic.SetHttpClient(client),
-	)
 }
